requirements.txt
Django>=5.0
django-ninja>=1.1
python-dotenv>=1.0
PyJWT>=2.8